Gut doubleclick.js
